IMPLEMENT SAVE TO FILE FUNCTIONALITY
Modify Observer interface and SaveToFile class to accept Map of
user as a parameter.
Implement Subject interface to UserServlet
Delete HelloWorldServlet servlet as it is not needed any more and has to be
modified every time changes are made to classes and interfaces shared
with UserServlet servlet to avoid errors in project.